Change function type and fix the issue of remainder index of the array.
1. Change the return type of functions from int to long.
2. When the length of the array is not divisible, the summation error is
solved.

+++ branch_prediction_array_summation/sum_array.h
@@ -7,8 +7,8 @@
 
 
 /* functions */
-int sum_array_simplest(int * array, int length);
-int sum_array_unroll(int * array, int length);
-int sum_array_duffs_device(int * array, int length);
+long sum_array_simplest(int * array, int length);
+long sum_array_unroll(int * array, int length);
+long sum_array_duffs_device(int * array, int length);
 
 #endif
